On login, only consider users if they are in the User default queryset.
So for example, if a Facebook Account is already associated with a user, but that user is for some reason not in the default queryset (say he belongs to a different "virtual site"), then a new user is created with a new Facebook-Assocation.

To make this work, the unique index on UserSocialAuth must be extended to allow the same Social account to be associated with multiple different users.